requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
groq>=0.13.0
PyPDF2==3.0.1
sentence-transformers==3.0.0
faiss-cpu==1.8.0
pydantic==2.9.0
numpy==1.26.4
